Rename --noscrape to --no-scrape to match other arguments
