+++ .idea/runConfigurations/AgTern__No_Scrape_.xml
@@ -13,7 +13,7 @@
     <option name="ADD_SOURCE_ROOTS" value="true" />
     <EXTENSION ID="PythonCoverageRunConfigurationExtension" runner="coverage.py" />
     <option name="SCRIPT_NAME" value="agtern" />
-    <option name="PARAMETERS" value="--noscrape" />
+    <option name="PARAMETERS" value="--no-scrape" />
     <option name="SHOW_COMMAND_LINE" value="false" />
     <option name="EMULATE_TERMINAL" value="false" />
     <option name="MODULE_MODE" value="true" />
